[iphone] Divide file `@/phone/Phone` from `@/phone`
